Added image controller, service and repository code, and removed the … (#14)
* First commit

* created dto and yaml

* Addressed review comments

* Addressed review comments

* Added controller and services and established relation between them

* Added product repository, service and controller code

* Changed to PagingAndSortingRepository and removed Optional for getting products by category

* Added image controller, service and repository code, and removed the image description feild

* Added cascading option to the parent table

* Resolved merge conflicts

* Added response entity to service class

* Addressed review comments

* Addressed review comments

+++ Product Management Service/Code/product-service/src/main/java/com/shopkeeper/productservice/service/ProductService.java
@@ -76,8 +76,11 @@ public class ProductService {
   }
 
   public ResponseEntity<?> deleteProductById(Integer id) {
-    productRepository.deleteById(id);
-    return new ResponseEntity<>("Deleted", HttpStatus.NO_CONTENT);
+    if (productRepository.existsById(id)) {
+      productRepository.deleteById(id);
+      return new ResponseEntity<>("Deleted", HttpStatus.NO_CONTENT);
+    }
+    return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
   }
 
   public ResponseEntity<?> deleteAllProducts() {
